refactor(addAuthPassword): move registration page logic flow to home page, replacing duplicate auth logic. Remove registration route.
